rename files to CamelCase
