+++ portal/code/tractus-x-portal/src/components/home.tsx
@@ -35,9 +35,9 @@ import { NewSemanticModel } from './semantics/NewSemanticModel';
 import SemanticHub from './semantics/SemanticHub';
 import SemanticModelDetail from './semantics/SemanticModelDetail';
 import Aspect from './aspect';
-import DigitalTwins from './digitaltwins/digitaltwins';
+import DigitalTwins from './digitaltwins/DigitalTwins';
 import Onboarding from './onboarding';
-import { DigitalTwinDetail } from './digitaltwins/digitaltwindetail';
+import { DigitalTwinDetail } from './digitaltwins/DigitalTwinDetail';
 
 const navStyles: Partial<INavStyles> = {
   root: {
@@ -230,7 +230,7 @@ class Home extends React.Component<RouteComponentProps> {
               <Route path='/home/newsemanticmodel' component={(props) => <NewSemanticModel {...props} />} />
               <Route path='/home/semanticmodel/:id' component={(props) => <SemanticModelDetail {...props} />} />
               <Route path='/home/aspect/:connector' component={(props) => <Aspect {...props} />} />
-              <Route path='/home/digitaltwins' component={(props) => <DigitalTwins {...props} />} />
+              <Route path='/home/DigitalTwins' component={(props) => <DigitalTwins {...props} />} />
               <Route path='/home/digitaltwin/:id' component={(props) => <DigitalTwinDetail {...props} />} />
               <Route path='/home/developerhub' component={(props) => <DeveloperHub {...props} />} />
               <Route path='/home/appdetail/:id' component={(props) => <AppDetail {...props} />} />
